Move likelihood tests to their respective models
